pyproject.toml
[tool.poetry]
name = "PEZ-ServerMonitor"
version = "0.1.0"
description = ""
authors = ["Timothy Bryant <timothybryant3@gmail.com>"]
readme = "README.md"
classifiers = ["Private :: Do not Upload"]

[tool.poetry.dependencies]
python = "^3.10"
distro = "^1.8.0"
netifaces = "^0.11.0"
psutil = "^5.9.5"
tabulate = "^0.9.0"
ping3 = "^4.0.4"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
sourcery = "^1.0.9"
pytest = "^7.2.2"
black = "^23.3.0"
coverage = "^7.2.6"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.pytest.ini_options]
# pythonpath = [
#   ".", "src"
# ]